+++ .jenkins/Jenkinsfile.groovy
@@ -40,6 +40,7 @@ node('compose-unity') {
 
 unityPackagePipeline {
     PACKAGE_LOCATION = '.jenkins/fixtures/unity-package'
+    PACKAGE_ID = 'net.slothsoft.jenkins-unity.integration'
     PACKAGE_BRANCH = 'main'
     TEST_CHANGELOG = true
     TEST_FORMATTING = true