+++ vars/deleteUnityProject.groovy
@@ -1,0 +1,10 @@
+import net.slothsoft.jenkins.unity.InstalledUnityPackage
+
+void call(InstalledUnityPackage installedPackage) {
+    if (!installedPackage) {
+        return
+    }
+    dir(installedPackage.workDirectory) {
+        deleteDir()
+    }
+}